+++ src/containers/questionnaire/utils/index.js
@@ -9,7 +9,7 @@ import {
 import { Constants } from 'lattice';
 import { DataProcessingUtils } from 'lattice-fabricate';
 
-import { QUESTION } from '../../../core/edm/constants/CollectionTemplateNames';
+import { QUESTION } from '../../../core/edm/constants/EntityTemplateNames';
 import { PROPERTY_TYPE_FQNS } from '../../../core/edm/constants/FullyQualifiedNames';
 
 const {
