+++ src/core/edm/constants/EntityTemplateNames.js
