+++ src/utils/FormUtils.js
@@ -8,7 +8,7 @@ import {
 } from 'immutable';
 import { DataProcessingUtils } from 'lattice-fabricate';
 
-import { PARTICIPANTS } from '../core/edm/constants/CollectionTemplateNames';
+import { PARTICIPANTS } from '../core/edm/constants/EntityTemplateNames';
 import { PROPERTY_TYPE_FQNS } from '../core/edm/constants/FullyQualifiedNames';
 
 const { getEntityAddressKey, getPageSectionKey, parseEntityAddressKey } = DataProcessingUtils;
